Major modifications to the maf_read() function so that it no longer has to call maf_peek().  Now it builds the list of sequences, the coordinate map, and the set of redundant sites as it goes through the MAF file.
This requires that the MAF is sorted with respect to the reference sequence.
If it is not, an error is reported (but execution does not stop).

A new function called maf_read_cats() was also added which takes a
cats_to_do list and only reads sites with those categories, which can
save lots of time on big MAF files.  maf_read() is now just a wrapper which
calls maf_read_cats with a NULL category list.


git-svn-id: svn+ssh://nextgen.cshl.edu/usr/local/svnroot/phast/trunk@1612 3e584e52-61a1-4911-b546-322c73959ec4