maf_parse: Allow maf blocks with duplicate srcNames.
This happens in large alignments, such as in the Zoonomia alignments
of 241 placental mammals.

There is a potential problem in allowing duplicate srcNames, if the
srcName is also in a list of sequences to include/exclude, reorder,
or mask, then this  might not be handled correctly. However, this
problem also exists when just the species name component appears more
than once, and is also in a list of sequences to include/exclude,
reorder, or mask.